v2.11.2
[*] restructuring of the build system
[*] restructuring of the rpm package generator
[*] updated INSTALL file
[*] some text edits everywhere
[*] reorganized header files
