+++ LICENSE
@@ -1,8 +1,9 @@
-Copyright MMXXII Amirul I
+Copyright MMXXV Amirul I
 
 Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+you may not use this library/project and all its assets except
+in compliance with the License. You may obtain a copy of the 
+License at
 
     http://www.apache.org/licenses/LICENSE-2.0
