+++ cmake/definitions.cmake
@@ -7,14 +7,13 @@ else()
 	set(LIB_SUFFIX "")
 endif()
 
-function(removeDuplicateSubstring stringIn stringOut)
-	separate_arguments(stringIn)
-	list(REMOVE_DUPLICATES stringIn)
-	string(REPLACE ";" " " stringIn "${stringIn}")
-	set(${stringOut} "${stringIn}" PARENT_SCOPE)
-endfunction()
-
 # configure_file(${CMAKE_CURRENT_SOURCE_DIR}/extras/axon.h.in ${CMAKE_CURRENT_SOURCE_DIR}/include/axon.h)
 # configure_file(${CMAKE_CURRENT_SOURCE_DIR}/include/axon/defines.h.in ${CMAKE_CURRENT_SOURCE_DIR}/include/axon/defines.h)
 
 # try_compile(HAS_STD_VARIANT ${CMAKE_BINARY_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/test/try_compile_variant.cpp)
+
+# get_cmake_property(_variableNames VARIABLES)
+# list (SORT _variableNames)
+# foreach (_variableName ${_variableNames})
+#     message(STATUS "${_variableName}=${${_variableName}}")
+# endforeach()
