+++ cmake/packaging.cmake
@@ -26,8 +26,10 @@ set(CPACK_PACKAGE_URL "https://github.com/thegeekyboy/axon")
 set(CPACK_RPM_PACKAGE_LICENSE "Apache-2.0")
 set(CPACK_RPM_PACKAGE_GROUP "Applications/Productivity")
 
+string (REGEX MATCH "\\.el[1-9]" OS_VERSION_PREFIX ${CMAKE_SYSTEM})
+
 set(CPACK_PACKAGING_INSTALL_PREFIX ${CMAKE_INSTALL_PREFIX})
-set(CPACK_PACKAGE_FILE_NAME "${CPACK_PACKAGE_NAME}-${CPACK_PACKAGE_VERSION}-${CPACK_PACKAGE_RELEASE}.${CMAKE_SYSTEM_PROCESSOR}")
+set(CPACK_PACKAGE_FILE_NAME "${CPACK_PACKAGE_NAME}-${CPACK_PACKAGE_VERSION}-${CPACK_PACKAGE_RELEASE}${OS_VERSION_PREFIX}.${CMAKE_SYSTEM_PROCESSOR}")
 set(CPACK_RPM_PACKAGE_RELOCATABLE FALSE)
 set(CPACK_RPM_PACKAGE_AUTOREQPROV " no")
