+++ examples/counter.cpp
@@ -1,3 +1,4 @@
+#include <iostream>
 #include <thread>
 #include <signal.h>
