+++ examples/mail.cpp
@@ -1,3 +1,5 @@
+#include <iostream>
+
 #include <axon.h>
 #include <axon/mail.h>
