+++ examples/rabbitmq.cpp
@@ -1,3 +1,4 @@
+#include <iostream>
 #include <functional>
 #include <csignal>
