+++ include/axon.h
@@ -2,14 +2,7 @@
 #define AXON_MASTER_H_
 
 #include <string>
-#include <utility>
-#include <sstream>
-#include <array>
-#include <vector>
-#include <functional>
-#include <chrono>
-#include <iomanip>
-
+#include <cstring>
 #include <sys/stat.h>
 
 #define MAXBUF 2097152 //1048576
@@ -31,6 +24,7 @@
 #endif
 
 #if DEBUG == 1
+	#include <mutex>
 	#define DBGPRN(...) RAWDBG(AXON_DEBUG, __VA_ARGS__)
 	#define INFPRN(...) RAWDBG(AXON_INFO, __VA_ARGS__)
 	#define NOTPRN(...) RAWDBG(AXON_NOTICE, __VA_ARGS__)
@@ -52,6 +46,8 @@ namespace axon
 	typedef int flags_t;
 	typedef int proto_t;
 	typedef int auth_t;
+
+	std::string version();
 
 	struct flags {
 
@@ -101,186 +97,41 @@ namespace axon
 		struct stat st;
 	};
 
-	struct licensekey
-	{
-		char version[8];
-		char key[64];
-		char validity[64];
-		int expiredate;
-	};
-
 	class exception : public std::exception {
 
 		char _what[4096];
-		std::string _msg;
-		std::string message;
+		char _msg[3072];
 
 		std::string make_message(const char* filename, int linenum, const std::string& source, const std::string& message)
 		{
-			std::stringstream s;
+			snprintf(_what, 4095, "%s: %d => %s(): %s", filename, linenum, source.c_str(), message.c_str());
+			strncpy(_msg, message.c_str(), 3071);
 
-			s<<filename<<":"<<linenum<<" => "<<source<<"(): "<<message<<std::endl;
-			_msg = message;
-
-			return s.str();
+			return _what;
 		};
 
 		public:
-		exception(): exception("none", 0, "axon::error", "no error message") { };
+		exception() = delete;
 		exception(const std::string& message): exception("none", 0, "axon:error", std::move(message)) { };
 		exception(const std::string& source, const std::string& message): exception("none", 0, std::move(source), std::move(message)) { };
-		exception(const char* filename, int linenum, const std::string& source, const std::string& message): message(make_message(filename, linenum, source, message)) { };
+		exception(const char* filename, int linenum, const std::string& source, const std::string& message) { make_message(filename, linenum, source, message); };
 		template<typename... Args> exception(const char* filename, int linenum, const std::string& source, const std::string& format, Args... args) {
 			char output[4096];
 			snprintf(output, 4096, format.c_str(), args...);
-			message = make_message(filename, linenum, source, output);
+			make_message(filename, linenum, source, output);
 		}
 
 		~exception() throw() {};
 
 		virtual const char* what() const throw () {
 
-			return message.c_str();
+			return _what;
 		};
 
 		virtual const char* msg() const throw () {
 
-			return _msg.c_str();
+			return _msg;
 		};
-	};
-
-	template<std::size_t N, class R, class...Args>
-	struct lambda
-	{
-		static std::array< std::function<R(Args&&...)>, N >& table()
-		{
-			static std::array< std::function<R(Args&&...)>, N > arr;
-			return arr;
-		}
-
-		template<std::size_t I>
-		static R call( Args...args )
-		{
-			return table()[I]( std::forward<Args>(args)... );
-		}
-
-		using sig = R(Args...);
-
-		template<std::size_t I=N-1>
-		static sig* make(std::function<R(Args&&...)> f)
-		{
-			if(!table()[I])
-			{
-				table()[I]=f;
-				return &call<I>;
-			}
-
-			if(I==0) return nullptr;
-
-			return make< (I-1)%N >(f);
-		}
-
-		template<std::size_t I=N-1>
-		static void recycle( sig* f )
-		{
-			if (f==call<I>)
-			{
-				table()[I]={};
-				return;
-			}
-
-			if (I==0) return;
-
-			recycle< (I-1)%N >(f);
-		}
-	};
-
-	struct timer {
-
-		std::chrono::time_point<std::chrono::high_resolution_clock> _start, _end;
-		std::vector<long> _laps;
-		std::string _name;
-
-		timer(const char *name):_name(name)
-		{
-			_start = std::chrono::high_resolution_clock::now();
-		}
-
-		timer(std::string &name):_name(name)
-		{
-			_start = std::chrono::high_resolution_clock::now();
-		}
-
-		~timer()
-		{
-			_end = std::chrono::high_resolution_clock::now();
-			auto microseconds = std::chrono::duration_cast<std::chrono::microseconds>(_end - _start);
-			INFPRN("%s ran for %ldμs", _name.c_str(), microseconds.count());
-		}
-
-		long now()
-		{
-			_end = std::chrono::high_resolution_clock::now();
-			std::chrono::microseconds microseconds = std::chrono::duration_cast<std::chrono::microseconds>(_end - _start);
-			return microseconds.count();
-		}
-
-		void reset()
-		{
-			_start = std::chrono::high_resolution_clock::now();
-			_laps.clear();
-		}
-
-		void lap()
-		{
-			std::chrono::time_point<std::chrono::high_resolution_clock> _temp = std::chrono::high_resolution_clock::now();
-			std::chrono::microseconds microseconds = std::chrono::duration_cast<std::chrono::microseconds>(_temp - _start);
-			_laps.push_back(microseconds.count());
-		}
-
-		template <
-			class result_t   = std::chrono::milliseconds,
-			class clock_t    = std::chrono::steady_clock,
-			class duration_t = std::chrono::milliseconds
-		>
-		auto since(std::chrono::time_point<clock_t, duration_t> const& start)
-		{
-			return std::chrono::duration_cast<result_t>(clock_t::now() - start);
-		}
-
-		static std::string iso8601()
-		{
-			std::stringstream ss;
-
-			const auto current_time_point {std::chrono::system_clock::now()};
-			const auto current_time_since_epoch {current_time_point.time_since_epoch()};
-			const auto current_milliseconds {std::chrono::duration_cast<std::chrono::milliseconds> (current_time_since_epoch).count() % 1000};
-
-			std::time_t t = std::chrono::system_clock::to_time_t(current_time_point);
-			ss<<std::put_time(std::localtime(&t), "%FT%T")<<"."<<std::setw(3)<<std::setfill('0')<<current_milliseconds;
-
-			return ss.str();
-		}
-
-		static long epoch()
-		{
-			return std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
-		}
-
-		static long diff(const long ep)
-		{
-			return std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count() - ep;
-		}
-
-		template <
-			class result_t   = std::chrono::milliseconds,
-			class clock_t    = std::chrono::steady_clock,
-			class duration_t = std::chrono::milliseconds
-		>
-		static auto diff(const std::chrono::time_point<std::chrono::high_resolution_clock> ep)
-		{
-			return std::chrono::duration_cast<result_t>(clock_t::now() - ep);
-		}
 	};
 }
