+++ include/axon/ldap.h
@@ -4,6 +4,9 @@
 #define STRINGIZE(x) STRINGIZE_A(x)
 #define STRINGIZE_A(x) #x
 
+#include <memory>
+#include <algorithm>
+
 #include STRINGIZE(LDAP_INCLUDEDIR/ldap.h)
 #include <sasl/sasl.h>
 
@@ -39,7 +42,7 @@ namespace axon {
 
 			struct error {
 
-				std::stringstream message;
+				std::string message;
 				bool isError;
 
 				error(LDAP *ldap, int rc) {
@@ -47,14 +50,14 @@ namespace axon {
 					isError = false;
 					if (rc != LDAP_SUCCESS) {
 						ldap_get_option(ldap, LDAP_OPT_DIAGNOSTIC_MESSAGE, (void*)&msg);
-						message<<ldap_err2string(rc)<<" - "<<msg;
+						message= std::string(ldap_err2string(rc))+" - "+msg;
 						ldap_memfree(msg);
 						isError = true;
 					}
 					else
-						message<<"no error";
+						message="no error";
 				}
-				std::string get() { return message.str(); }
+				std::string get() { return message; }
 				operator bool() { return isError; }
 			};
