+++ include/axon/scylladb.h
@@ -62,7 +62,6 @@ namespace axon
 					size_t sz;
 
 					cass_iterator_get_meta_field_name(xd, &name, &sz);
-					std::cout<<name<<std::endl;
 				}
 
 				cass_iterator_free(xd);
