+++ include/axon/util.h
@@ -1,21 +1,102 @@
 #ifndef AXON_UTIL_H_
 #define AXON_UTIL_H_
 
-#include <iostream>
-#include <variant>
+#include <vector>
 #include <cstdarg>
-
-#include <boost/uuid/uuid.hpp>
-#include <boost/uuid/uuid_generators.hpp>
-#include <boost/uuid/uuid_io.hpp>
-
-#include <curl/curl.h>
+#include <chrono>
+#include <iomanip>
 
 #include <axon.h>
 
 namespace axon
 {
-	std::string version();
+	struct timer {
+
+		std::chrono::time_point<std::chrono::high_resolution_clock> _start, _end;
+		std::vector<long> _laps;
+		std::string _name;
+
+		timer(const char *name):_name(name)
+		{
+			_start = std::chrono::high_resolution_clock::now();
+		}
+
+		timer(std::string &name):_name(name)
+		{
+			_start = std::chrono::high_resolution_clock::now();
+		}
+
+		~timer()
+		{
+			_end = std::chrono::high_resolution_clock::now();
+			auto microseconds = std::chrono::duration_cast<std::chrono::microseconds>(_end - _start);
+			INFPRN("%s ran for %ldμs", _name.c_str(), microseconds.count());
+		}
+
+		long now()
+		{
+			_end = std::chrono::high_resolution_clock::now();
+			std::chrono::microseconds microseconds = std::chrono::duration_cast<std::chrono::microseconds>(_end - _start);
+			return microseconds.count();
+		}
+
+		void reset()
+		{
+			_start = std::chrono::high_resolution_clock::now();
+			_laps.clear();
+		}
+
+		void lap()
+		{
+			std::chrono::time_point<std::chrono::high_resolution_clock> _temp = std::chrono::high_resolution_clock::now();
+			std::chrono::microseconds microseconds = std::chrono::duration_cast<std::chrono::microseconds>(_temp - _start);
+			_laps.push_back(microseconds.count());
+		}
+
+		template <
+			class result_t   = std::chrono::milliseconds,
+			class clock_t    = std::chrono::steady_clock,
+			class duration_t = std::chrono::milliseconds
+		>
+		auto since(std::chrono::time_point<clock_t, duration_t> const& start)
+		{
+			return std::chrono::duration_cast<result_t>(clock_t::now() - start);
+		}
+
+		static std::string iso8601()
+		{
+			std::stringstream ss;
+
+			const auto current_time_point {std::chrono::system_clock::now()};
+			const auto current_time_since_epoch {current_time_point.time_since_epoch()};
+			const auto current_milliseconds {std::chrono::duration_cast<std::chrono::milliseconds> (current_time_since_epoch).count() % 1000};
+
+			std::time_t t = std::chrono::system_clock::to_time_t(current_time_point);
+			ss<<std::put_time(std::localtime(&t), "%FT%T")<<"."<<std::setw(3)<<std::setfill('0')<<current_milliseconds;
+
+			return ss.str();
+		}
+
+		static long epoch()
+		{
+			return std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+		}
+
+		static long diff(const long ep)
+		{
+			return std::chrono::duration_cast<std::chrono::seconds>(std::chrono::system_clock::now().time_since_epoch()).count() - ep;
+		}
+
+		template <
+			class result_t   = std::chrono::milliseconds,
+			class clock_t    = std::chrono::steady_clock,
+			class duration_t = std::chrono::milliseconds
+		>
+		static auto diff(const std::chrono::time_point<std::chrono::high_resolution_clock> ep)
+		{
+			return std::chrono::duration_cast<result_t>(clock_t::now() - ep);
+		}
+	};
 
 	namespace util
 	{
@@ -75,12 +156,6 @@ namespace axon
 
 			return cnt;
 		}
-
-		template <typename T, typename... Args> struct concatenator;
-		template <typename... Args0, typename... Args1>
-		struct concatenator<std::variant<Args0...>, Args1...> {
-			using type = std::variant<Args0..., Args1...>;
-		};
 	}
 }
