--- include/axon/version.h
@@ -1,1 +1,0 @@
-#define VERSION "2.10.3-5423"
