+++ src/CMakeLists.txt
@@ -1,39 +1,31 @@
 #
-target_sources(${PROJECT_NAME}
-	PRIVATE
-	${CMAKE_CURRENT_SOURCE_DIR}/util.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/log.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/md5.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/aes.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/dmi.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ftplist.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/config.cpp
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/util.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/log.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/md5.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/aes.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/dmi.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/ftplist.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/config.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/fstream.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/connection.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/nothing.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/file.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/ftp.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/ssh.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/s3.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/samba.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/hdfs.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/sqlite.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/scylladb.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/oracle.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/kafka.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/cqn.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/rabbit.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/kerberos.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/ldap.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/message.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/mml.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/socket.cpp)
+list(APPEND SRCFILES ${CMAKE_CURRENT_SOURCE_DIR}/mail.cpp)
 
-	${CMAKE_CURRENT_SOURCE_DIR}/fstream.cpp
-
-	${CMAKE_CURRENT_SOURCE_DIR}/connection.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/nothing.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/file.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ftp.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ssh.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/s3.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/samba.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/hdfs.cpp
-
-	${CMAKE_CURRENT_SOURCE_DIR}/sqlite.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/scylladb.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/oracle.cpp
-
-	${CMAKE_CURRENT_SOURCE_DIR}/kafka.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/cqn.cpp
-
-	${CMAKE_CURRENT_SOURCE_DIR}/rabbit.cpp
-
-	${CMAKE_CURRENT_SOURCE_DIR}/kerberos.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ldap.cpp
-
-	${CMAKE_CURRENT_SOURCE_DIR}/message.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/mml.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/socket.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/mail.cpp
-)
+target_sources(${PROJECT_NAME} PRIVATE ${SRCFILES})
