+++ src/cqn.cpp
@@ -121,18 +121,18 @@ namespace axon {
 
 				if (table_op & axon::database::change::allrows) {
 
-					if (table_op & axon::database::change::insert)
-						std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - insert"<<std::endl;
-					else if (table_op & axon::database::change::update)
-						std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - update"<<std::endl;
-					else if (table_op & axon::database::change::remove)
-						std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - remove"<<std::endl;
-					else if (table_op & axon::database::change::alter)
-						std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - alter"<<std::endl;
-					else if (table_op & axon::database::change::drop)
-						std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - drop"<<std::endl;
-					else if (table_op & axon::database::change::unknown)
-						std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - unknown"<<std::endl;
+					// if (table_op & axon::database::change::insert)
+					// 	std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - insert"<<std::endl;
+					// else if (table_op & axon::database::change::update)
+					// 	std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - update"<<std::endl;
+					// else if (table_op & axon::database::change::remove)
+					// 	std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - remove"<<std::endl;
+					// else if (table_op & axon::database::change::alter)
+					// 	std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - alter"<<std::endl;
+					// else if (table_op & axon::database::change::drop)
+					// 	std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - drop"<<std::endl;
+					// else if (table_op & axon::database::change::unknown)
+					// 	std::cout<<"table: "<<table_name<<" = OCI_OPCODE_ALLROWS <> "<<table_op<<" - unknown"<<std::endl;
 
 					continue;
 				}
@@ -172,16 +172,6 @@ namespace axon {
 		{
 			axon::stream::message_t data = {false, op, ch, table, rowid};
 			_pipe.push_back(std::move(data));
-			// _pipe.emplace_front(blah);
-
-			// std::string sql = "SELECT * FROM " + table + " WHERE ROWID = '" + rowid + "'";
-
-			// std::shared_ptr<axon::database::statement> stmt = std::make_shared<axon::database::statement>(*_context);
-			// stmt->prepare(sql);
-			// stmt->execute(axon::database::exec_type::select);
-
-			// std::unique_ptr<axon::database::resultset> rs = std::make_unique<axon::database::resultset>(&stmt);
-			// _topic.callback(op, ch, table, rowid, std::move(rs));
 			_topic.callback(op, ch, table, rowid, this);
 		}
 
@@ -220,7 +210,6 @@ namespace axon {
 			}
 		}
 
-		// std::string cqn::subscribe(std::string topic, std::function<void(axon::stream::recordset*)> interceptor)
 		std::string cqn::subscribe(std::string topic, cbfn interceptor)
 		{
 			axon::timer ctm(__PRETTY_FUNCTION__);
@@ -233,9 +222,8 @@ namespace axon {
 
 		std::string cqn::subscribe(std::string topic)
 		{
-			// return subscribe(topic, &test);
-			return subscribe(topic, [](axon::database::operation, axon::database::change, std::string, std::string rowid, void *) {
-				std::cout<<rowid<<std::endl;
+			return subscribe(topic, [](axon::database::operation, axon::database::change, std::string, std::string, void *) {
+				//TODO: do something here
 			});
 		}
