+++ src/rabbit.cpp
@@ -1,4 +1,5 @@
 #include <string>
+#include <cstring>
 #include <sstream>
 
 #include <axon/util.h>
