+++ src/scylladb.cpp
@@ -1,6 +1,7 @@
 #include <any>
 #include <regex>
 #include <cstdarg>
+#include <iostream>
 
 #include <axon.h>
 #include <axon/util.h>
@@ -149,10 +150,7 @@ namespace axon
 						throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "failed to bind at index: %d, driver: %s", index, cass_error_desc(ce));
 				}
 				else
-				{
-					std::cout<<element.type().name()<<std::endl;
 					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "unsupported data type %s at index: %d", element.type().name(), index);
-				}
 
 				index++;
 			}
