+++ src/util.cpp
@@ -1,3 +1,4 @@
+#include <iostream>
 #include <iomanip>
 #include <sstream>
 #include <string>
@@ -13,7 +14,7 @@
 #include <boost/filesystem.hpp>
 
 #include <axon.h>
-#include <axon/version.h>
+#include <version.h>
 #include <axon/util.h>
 #include <axon/md5.h>
 #include <axon/aes.h>
@@ -63,15 +64,6 @@ namespace axon
 			ltrim(s);
 		}
 
-		unsigned int toInt(char c)
-		{
-			if (c >= '0' && c <= '9') return c - '0';
-			if (c >= 'A' && c <= 'F') return 10 + c - 'A';
-			if (c >= 'a' && c <= 'f') return 10 + c - 'a';
-
-			return c;
-		}
-
 		unsigned long long bytestoull(const char *bcd, const size_t size)
 		{
 			std::stringstream ss;
@@ -97,18 +89,6 @@ namespace axon
 			}
 
 			return ss.str();
-		}
-
-
-		unsigned long getdate()
-		{
-			time_t rawtime;
-			struct tm *timeinfo;
-
-			rawtime = time (NULL);
-			timeinfo = localtime(&rawtime);
-
-			return ((timeinfo->tm_year+1900)*10000) + ((timeinfo->tm_mon+1)*100) + timeinfo->tm_mday;
 		}
 
 		char *trim(char *str)
@@ -363,110 +343,6 @@ namespace axon
 				magic_close(mcookie);
 			}
 			return { mimetype, encoding };
-		}
-
-		int license(struct licensekey *master)
-		{
-			md5_state_t state;
-			md5_byte_t digest[16];
-			uint8_t hash[16], in[16], out[16];
-			char hex_output[16*2 + 1], output[1024], seed[2048];
-			int cnt, cdn;
-			FILE *fp;
-
-			char s1[1024] = {0x2f, 0x75, 0x73, 0x72, 0x2f, 0x62, 0x69, 0x6e, 0x2f, 0x68, 0x61, 0x6c, 0x2d, 0x67, 0x65, 0x74, 0x2d, 0x70, 0x72, 0x6f, 0x70, 0x65, 0x72, 0x74, 0x79, 0x20, 0x2d, 0x2d, 0x75, 0x64, 0x69, 0x20, 0x2f, 0x6f, 0x72, 0x67, 0x2f, 0x66, 0x72, 0x65, 0x65, 0x64, 0x65, 0x73, 0x6b, 0x74, 0x6f, 0x70, 0x2f, 0x48, 0x61, 0x6c, 0x2f, 0x64, 0x65, 0x76, 0x69, 0x63, 0x65, 0x73, 0x2f, 0x63, 0x6f, 0x6d, 0x70, 0x75, 0x74, 0x65, 0x72, 0x20, 0x2d, 0x2d, 0x6b, 0x65, 0x79, 0x20, 0x73, 0x79, 0x73, 0x74, 0x65, 0x6d, 0x2e, 0x68, 0x61, 0x72, 0x64, 0x77, 0x61, 0x72, 0x65, 0x2e, 0x75, 0x75, 0x69, 0x64, 0x20, 0x32, 0x3e, 0x2f, 0x64, 0x65, 0x76, 0x2f, 0x6e, 0x75, 0x6c, 0x6c, 0x00};
-			char s2[1024] = {0x2f, 0x73, 0x62, 0x69, 0x6e, 0x2f, 0x69, 0x66, 0x63, 0x6f, 0x6e, 0x66, 0x69, 0x67, 0x20, 0x2d, 0x61, 0x20, 0x7c, 0x20, 0x67, 0x72, 0x65, 0x70, 0x20, 0x2d, 0x45, 0x20, 0x22, 0x28, 0x65, 0x6d, 0x31, 0x7c, 0x65, 0x74, 0x68, 0x30, 0x29, 0x22, 0x20, 0x7c, 0x20, 0x74, 0x72, 0x20, 0x2d, 0x73, 0x20, 0x27, 0x20, 0x27, 0x20, 0x7c, 0x20, 0x63, 0x75, 0x74, 0x20, 0x2d, 0x64, 0x27, 0x20, 0x27, 0x20, 0x2d, 0x66, 0x35, 0x00};
-			//char s2[1024] = {0x2f, 0x62, 0x69, 0x6e, 0x2f, 0x63, 0x61, 0x74, 0x20, 0x2f, 0x73, 0x79, 0x73, 0x2f, 0x63, 0x6c, 0x61, 0x73, 0x73, 0x2f, 0x6e, 0x65, 0x74, 0x2f, 0x65, 0x6d, 0x31, 0x2f, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x20, 0x32, 0x3e, 0x2f, 0x64, 0x65, 0x76, 0x2f, 0x6e, 0x75, 0x6c, 0x6c, 0x00}; //em0
-			//char s2[1024] = {0x2f, 0x62, 0x69, 0x6e, 0x2f, 0x63, 0x61, 0x74, 0x20, 0x2f, 0x73, 0x79, 0x73, 0x2f, 0x63, 0x6c, 0x61, 0x73, 0x73, 0x2f, 0x6e, 0x65, 0x74, 0x2f, 0x65, 0x74, 0x68, 0x30, 0x2f, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x20, 0x32, 0x3e, 0x2f, 0x64, 0x65, 0x76, 0x2f, 0x6e, 0x75, 0x6c, 0x6c, 0x00}; //eth0
-
-			if ((fp = popen(s1, "r")) == NULL)
-				return false;
-
-			bzero(output, 1024);
-
-			if (fgets(output, sizeof(output)-1, fp) == NULL)
-				return false;
-
-			pclose(fp);
-
-			strcpy(seed, output);
-			seed[strlen(seed)-1] = 0;
-
-			if ((fp = popen(s2, "r")) == NULL)
-				return false;
-
-			bzero(output, 1024);
-
-			if (fgets(output, sizeof(output)-1, fp) == NULL)
-				return false;
-
-			pclose(fp);
-
-			strcat(seed, output);
-			seed[strlen(seed)-1] = 0;
-
-			md5_init(&state);
-			md5_append(&state, (const md5_byte_t *) seed, strlen(seed));
-			md5_finish(&state, digest);
-
-			for (cnt = 0, cdn = 15; cnt < 16; cnt++, cdn--)
-			{
-				sprintf(hex_output + cnt * 2, "%02x", digest[cnt]);
-				hash[cnt] = digest[cdn];
-			}
-		#ifdef SHOWHASH
-
-			printf("HashKey>%s\n", hex_output);
-
-		#endif
-
-			if (strcmp(master->key, hex_output) != 0)
-			{
-				//printlog("FATL", "Overmind - The license key is wrong, please collect correct licence key.");
-				return false;
-			}
-
-			for (cnt = 0; cnt < 16; cnt++)
-				in[cnt] = 16 * toInt(master->validity[cnt*2]) + toInt(master->validity[(cnt*2)+1]);
-
-		#ifdef SHOWHASH
-
-			printf("NewVali>");
-
-			uint8_t test[16] = {0x20, 0x17, 0x06, 0x30, 0X46, 0x49, 0x4c, 0x45, 0x42, 0x4f, 0x54, 0x56, 0x30, 0x2e, 0x31, 0x30};
-			AES128_ECB_encrypt(test, hash, out);
-			for (cnt = 0; cnt < 16; cnt++)
-				printf("%02x", out[cnt]);
-				printf("\n");
-
-		#endif
-
-			AES128_ECB_decrypt(in, hash, out);
-
-			int chksum = false;
-
-			if (out[4] == 0x46 && out[5] == 0x49 && out[6] == 0x4c && out[7] == 0x45 && out[8] == 0x42 && out[9] == 0x4f && out[10] == 0x54 && out[11] == 0x56 && out[12] == 0x30 && out[13] == 0x2e && out[14] == 0x31 && out[15] == 0x30)
-				chksum = true;
-
-			if (!chksum)
-			{
-				//printlog("FATL", "Overmind - Data is corrupted. Please contact sisscons@yahoo.com");
-				return false;
-			}
-
-			unsigned long expiredate = ((((out[0]/16)*10)+(out[0]%16))*1000000) + ((((out[1]/16)*10)+(out[1]%16))*10000) + ((((out[2]/16)*10)+(out[2]%16))*100) + (((out[3]/16)*10)+(out[3]%16));
-
-			//printf("%d > %d\n", getdate(), expiredate);
-			if (getdate() > expiredate)
-			{
-				//printlog("FATL", "Overmind - Your licence has expired, kindly get or renew your licence");
-				return false;
-			} else {
-				//printlog("INFO", "Overmind - Your license will expire on: %02d-%02d-20%02d", (((out[3]/16)*10)+(out[3]%16)), (((out[2]/16)*10)+(out[2]%16)), (((out[1]/16)*10)+(out[1]%16)));
-			}
-
-			master->expiredate = expiredate;
-			return true;
 		}
 
 		bool execmd(const char *cmd)
